[Patch] Changed tabIndex from steps to parent in wizard component
