+++ projects/dxc-ngx-cdk/src/lib/dxc-wizard/dxc-wizard-step/dxc-wizard-step.component.html
@@ -9,7 +9,7 @@
   }"
   [disabled]="disabled"
   (click)="!disabled ? handleStepClick() : ''"
-  [tabindex]="tabIndexValue"
+  [tabindex]="disabled ? -1 : tabIndexValue"
 >
   <div class="stepHeader">
     <div class="iconContainer">
