+++ projects/dxc-ngx-cdk/src/lib/dxc-wizard/services/wizard.service.ts
@@ -14,6 +14,7 @@ export class WizardService {
     undefined
   );
   public mode: BehaviorSubject<string> = new BehaviorSubject(undefined);
+  public tabIndexValue: BehaviorSubject<number> = new BehaviorSubject(0);
 
   constructor() {
     this.innerCurrentStep.subscribe((newCurrent) => {
